+++ aixm_geo/settings.py
@@ -1,0 +1,9 @@
+# Declare AIXM namespace constant
+
+NAMESPACES = {
+    'xlink': "http://www.w3.org/1999/xlink",
+    'gml': "http://www.opengis.net/gml/3.2", 'aixm': "http://www.aixm.aero/schema/5.1",
+    'nats': "http://www.aixm.aero/schema/5.1/extensions/NATS/eSDO",
+    'message': "http://www.aixm.aero/schema/5.1/message",
+    'xsi': "http://www.w3.org/2001/XMLSchema-instance"
+}
